Wiring update on the architecture
Some wires are missing on the modules (verify Todos). Refer to comments on how to follow wiring convention. There is a more "clean" way to wire this, however it is harder to debug on the long run, and the convention makes it easier to debug on both code and GTK Wave without doing much documentation